Try some more tuning
* Log the state of apache processes when we restart so we can continue
  to understand the problem
* Restart slightly less often
* Update MaxConnectionsPerChild to 50, which is probably a bit more
  reasonable than 5
* Update server limit (see below)
* Turn off KeepAlive (see below)

From scale-infra, explanation on the last two above.

```
This is a confirmed bug in Apache with no configuration workaround:

https://bz.apache.org/bugzilla/show_bug.cgi?id=53555

There's a pretty good description of the details, which match our behavior here:

https://serverfault.com/questions/516373/what-is-the-meaning-of-ah00485-scoreboard-is-full-not-at-maxrequestworkers

One thing that some people seem to have had success with is turning keepalive off.

ServerLimit, for the event MPM (what we use), the recommendation is:

With event, increase this directive if the process number defined by your MaxRequestWorkers and ThreadsPerChild settings, plus the number of gracefully shutting down processes, is more than 16 server processes (default).

Our MaxRequestWorkers is 50. We don't define ThreadsPerChild which defaults to 25. So if I read that correctly (and I haven't tuned Apache for a living in a looonnnggg time), we want something like 50+25=75 plus some more for shutting down processes, so like... 80? I can prep a diff for that plus keepalive and see how that goes.
```

All continuing https://github.com/socallinuxexpo/scale-drupal/issues/79

Signed-off-by: Phil Dibowitz <phil@ipom.com>