Restart apache slightly less often
Part of https://github.com/socallinuxexpo/scale-chef/pull/293 was
supposed to drop restarts to hourly, but a rebase dropped that because
those lines were changed by Ilan to fix a bug I had introduced.

This drops us down to hourly. While we may run into issues, if we do
at least now we have logging, and at worst will be down for less than an
hour.

Signed-off-by: Phil Dibowitz <phil@ipom.com>